+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/CombinationGraph/BasicCombinationGraphs/PairedResultWithScatter.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import CombinationGraph from '@cerner/terra-graphs/lib/components/Combination/CombinationGraph';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import lineAndPairedResultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Combination/lineAndPairedResultConfig';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/PairedResultWithScatterData';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(lineAndPairedResultConfig('#pairedResultWithScatter'));
+const dataset = utils.deepClone(data);
+
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <CombinationGraph graphID="pairedResultWithScatter" graphConfig={graphConfig} dataset={dataset} />
+  </React.Fragment>
+);