+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -10,3 +10,4 @@
   * Added data objects and examples for Gantt chart.
   * Added data objects and examples for Timeline graph.
   * Added data objects and examples for Pie chart.
+  * Added data objects and examples for Combination graph.